website/admins/announcement_admin.py
from django.contrib import admin


class AnnouncementAdmin(admin.ModelAdmin):

    class Media:
        js = [
            'js/tinymce/tinymce.min.js',
            'js/tinymce_settings.js',
        ]
